Refactor test signers
